Feature/url verification (#3)
* added a method which finds the first URL in a given String

* formatting, fixed chain of command pattern, URL check now works for multiple URLs in one message

* Added URLVerificationSmsHandler to the main flow

* Links are now persisted in the database

* ThreatType and ConfidenceLevel enums are now properly persisted in the database

* updated README.md

* Added a method comparing threat confidence level with a threshold

* typo fix

+++ src/main/java/pl/jacekk/phishingdetector/repository/ContractRepository.java
@@ -5,6 +5,6 @@ import pl.jacekk.phishingdetector.entity.ContractEntity;
 
 import java.util.Optional;
 
-public interface ContractRepository extends CrudRepository<ContractEntity, Long> {
+public interface ContractRepository extends CrudRepository<ContractEntity, Integer> {
     Optional<ContractEntity> findByMsisdn(String msisdn);
 }
